requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar
Django == 3.2.1
django-cors-headers == 3.6.0
psycopg2 == 2.8.6
social-auth-app-django==4.0.0
djangorestframework == 3.12.2
djangorestframework-simplejwt == 4.4.0
django_simple_captcha == 0.5.13
drf-yasg == 1.20.0
baidu-aip == 2.2.18.0
django-filter == 2.4.0
face_recognition == 1.3.0
numpy == 1.19.4
pandas == 1.1.5
ruamel.yaml == 0.17.4
django-simpleui == 2021.5.11
django-import-export == 2.5.0